move all scatterplot utils into single file
